+++ grafana_backup/cli.py
@@ -1,16 +1,19 @@
 from grafana_backup.constants import (PKG_NAME, PKG_VERSION, JSON_CONFIG_PATH)
 from grafana_backup.save import main as save
 from grafana_backup.restore import main as restore
+from grafana_backup.delete import main as delete
 from grafana_backup.grafanaSettings import main as conf
 from docopt import docopt
-import os, sys
+import os
+import sys
 
 docstring = """
 {0} {1}
 
 Usage:
-    grafana-backup save [--config=<filename>] [--components=<folders,dashboards,datasources,alert-channels,organizations,users>] [--no-archive]
-    grafana-backup restore <archive_file> [--config=<filename>] [--components=<folders,dashboards,datasources,alert-channels,organizations,users>]
+    grafana-backup save [--config=<filename>] [--components=<folders,dashboards,datasources,alert-channels,organizations,users,snapshots,versions,annotations>] [--no-archive]
+    grafana-backup restore <archive_file> [--config=<filename>] [--components=<folders,dashboards,datasources,alert-channels,organizations,users,snapshots,annotations>]
+    grafana-backup delete [--config=<filename>] [--components=<folders,dashboards,datasources,alert-channels,snapshots,annotations>]
     grafana-backup [--config=<filename>]
     grafana-backup -h | --help
     grafana-backup --version
@@ -19,8 +22,8 @@ Options:
     -h --help                                                       Show this help message and exit
     --version                                                       Get version information and exit
     --config=<filename>                                             Override default configuration path
-    --components=<folders,dashboards,datasources,alert-channels,organizations,users>    Comma separated list of individual components to backup
-                                                                    rather than backing up all components by default
+    --components=<folders,dashboards,datasources,alert-channels,organizations,users,snapshots,versions,annotations>    Comma separated list of individual components to backup
+                                                                    rather than backing up all components by default. Versions can only be saved not restored.
     --no-archive                                                    Skip archive creation and do not delete unarchived files
                                                                     (used for troubleshooting purposes)
 """.format(PKG_NAME, PKG_VERSION)
@@ -44,6 +47,9 @@ def main():
         sys.exit()
     elif args.get('restore', None):
         restore(args, settings)
+        sys.exit()
+    elif args.get('delete', None):
+        delete(args, settings)
         sys.exit()
     else:
         print(docstring)