+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.110] - 2021-09-02
+
+### Added
+- Option to save and restore snapshots
+- Option to save dashboard versions (restore isn't really possible with the API)
+
+
 # [1.1.9] - 2021-06-27
 
 ### Added